requirements.txt
# --- Frameworks ---
Django==5.2.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1

# --- Conector do Banco de Dados ---
psycopg2-binary==2.9.10


# --- Servidor de Aplicação ---
gunicorn==23.0.0


# --- Utilitários ---
python-dotenv==1.1.1


# --- Outros ---
asgiref==3.9.1
django-cors-headers==4.7.0
packaging==25.0
PyJWT==2.10.1
sqlparse==0.5.3

# --- Manipulação de imagens ---
pillow==11.3.0

# --- Integração com AWS S3 ---
boto3==1.40.7
botocore==1.40.7
django-storages==1.14.6

s3transfer==0.13.1
jmespath==1.0.1
python-dateutil==2.9.0.post0
six==1.17.0
urllib3==2.5.0

black==25.1.0
click==8.2.1
mypy_extensions==1.1.0
pathspec==0.12.1
platformdirs==4.4.0

django-localflavor==5.0
phonenumbers==9.0.14
django-phonenumber-field==8.1.0
django-filter==25.1